Use single jest config
Thanks to @jest-environment doc annotations; fixes coverage reports
