--- packages/wrapper/jest.config.jsdom.js
@@ -1,7 +1,0 @@
-const config = require('next-redux-wrapper-configs/jest.config');
-module.exports = Object.assign({}, config, {
-    roots: ['<rootDir>/src'],
-    testPathIgnorePatterns: config.testPathIgnorePatterns.concat(['./es6', './lib', './types']),
-    testRegex: '/src/.*\\.spec\\.tsx?$',
-    testEnvironment: 'jsdom',
-});
